Added test for -no-alias-deps.
The test module Namespace.Nested now includes an alias for a submodule
Not_linked. Since Namespace.Nested is referenced, if -no-alias-deps is
not supplied, it will pull in Not_linked. Not_linked exits the test
program with status 1 as a side effect of initialization, in order to
detect this error.