Decoupled libraries from namespaces.
The namespace_lib tag is now parametric, and need not apply to a
namespace directory. namespace_lib(foo) causes all tagged files to be
placed into a library foo.cm[x]a. When applied to a directory, it
applies to all files and namespaces in that directory (including the
directory itself, if it is a namepsace), except to those that are
overridden by a deeper namespace_lib(bar) tag.

Files that are not tagged with namespace_lib, and are not in a
directory tagged with namespace_lib, but are in a namespace, are placed
into a special library _namespaces.cm[x]a. This is necessary for
achieving weak dependencies between these files and any executables
that are being built in the same project.

Libraries are generated at the top level of the source tree.

So, for example, if you have

  "src/": namespace_lib(mylib)
  "src/foo": namespace
  "src/foo/bar.ml": namespace_lib(myotherlib)

Then Foo and all its siblings and children, besides Foo.Bar, will be
placed in mylib.cm[x]a, while Foo.Bar will be placed in
myotherlib.cm[x]a.

Also:
- The modules in .mllib files are now sorted topologically according to
  the internal dependency relation. Lack of such sorting was probably a
  bug before, but it became especially acute during work on this
  commit.
- Now ignoring all files with name myocamlbuild.ml and setup.ml.